python-gesture-engine/requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
websockets>=12.0
numpy>=1.24.0
msgpack>=1.0.0
pyyaml>=6.0
